src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile src/requirements.in
#
apimapper==0.7.9
    # via -r src/requirements.in
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
brotli==1.1.0
    # via whitenoise
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
concurrent-log-handler==0.9.24
    # via -r src/requirements.in
defusedxml==0.7.1
    # via -r src/requirements.in
django==4.2.7
    # via
    #   -r src/requirements.in
    #   django-auth-ldap
    #   django-axes
    #   django-cas-ng
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-ranged-response
    #   django-redis
    #   django-simple-captcha
    #   djangorestframework
    #   drf-spectacular
    #   easy-thumbnails
django-auth-ldap==4.6.0
    # via -r src/requirements.in
django-axes[ipware]==6.1.1
    # via
    #   -r src/requirements.in
    #   django-axes
django-cas-ng==5.0.1
    # via -r src/requirements.in
django-cors-headers==4.3.0
    # via -r src/requirements.in
django-debug-toolbar==4.2.0
    # via -r src/requirements.in
django-environ==0.11.2
    # via -r src/requirements.in
django-filter==23.5
    # via -r src/requirements.in
django-ipware==5.0.2
    # via django-axes
django-mama-cas==2.5.0
    # via -r src/requirements.in
django-ranged-response==0.2.0
    # via django-simple-captcha
django-redis==5.4.0
    # via -r src/requirements.in
django-simple-captcha==0.5.20
    # via -r src/requirements.in
djangorestframework==3.14.0
    # via
    #   -r src/requirements.in
    #   drf-spectacular
djangorestframework-api-key==3.0.0
    # via -r src/requirements.in
drf-spectacular==0.27.0
    # via -r src/requirements.in
easy-thumbnails==2.8.5
    # via -r src/requirements.in
gevent==23.9.1
    # via -r src/requirements.in
greenlet==3.0.1
    # via gevent
gunicorn==21.2.0
    # via -r src/requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
isodate==0.6.1
    # via rdflib
jsonschema==4.20.0
    # via drf-spectacular
jsonschema-specifications==2023.11.2
    # via jsonschema
lxml==4.9.3
    # via python-cas
markdown==3.5.1
    # via apimapper
packaging==23.2
    # via
    #   build
    #   djangorestframework-api-key
    #   gunicorn
pillow==10.1.0
    # via
    #   -r src/requirements.in
    #   django-simple-captcha
    #   easy-thumbnails
pip-tools==7.3.0
    # via -r src/requirements.in
portalocker==2.8.2
    # via concurrent-log-handler
psutil==5.9.6
    # via rainbow-saddle
psycopg[binary]==3.1.12
    # via
    #   -r src/requirements.in
    #   psycopg
psycopg-binary==3.1.12
    # via psycopg
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.3.0
    # via python-ldap
pyparsing==3.1.1
    # via rdflib
pyproject-hooks==1.0.0
    # via build
python-cas==1.6.0
    # via django-cas-ng
python-ldap==3.4.3
    # via django-auth-ldap
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
rainbow-saddle==0.4.0
    # via -r src/requirements.in
rdflib==7.0.0
    # via skosmos-client
redis==5.0.1
    # via django-redis
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r src/requirements.in
    #   apimapper
    #   django-mama-cas
    #   python-cas
    #   skosmos-client
rpds-py==0.13.2
    # via
    #   jsonschema
    #   referencing
shortuuid==1.0.11
    # via -r src/requirements.in
six==1.16.0
    # via
    #   isodate
    #   python-cas
skosmos-client==0.3.0
    # via -r src/requirements.in
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.8.0
    # via psycopg
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.0.7
    # via requests
wheel==0.41.3
    # via pip-tools
whitenoise[brotli]==6.6.0
    # via
    #   -r src/requirements.in
    #   whitenoise
zope-event==5.0
    # via gevent
zope-interface==6.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
